requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
altair=4.2.2=pypi_0
attrs=22.2.0=pypi_0
blinker=1.5=pypi_0
ca-certificates=2023.01.10=haa95532_0
cachetools=5.3.0=pypi_0
certifi=2022.12.7=py39haa95532_0
charset-normalizer=3.0.1=pypi_0
click=8.1.3=pypi_0
colorama=0.4.6=pypi_0
decorator=5.1.1=pypi_0
entrypoints=0.4=pypi_0
gitdb=4.0.10=pypi_0
gitpython=3.1.30=pypi_0
idna=3.4=pypi_0
importlib-metadata=6.0.0=pypi_0
jinja2=3.1.2=pypi_0
jsonschema=4.17.3=pypi_0
markdown-it-py=2.1.0=pypi_0
markupsafe=2.1.2=pypi_0
mdurl=0.1.2=pypi_0
numpy=1.24.1=pypi_0
openssl=1.1.1s=h2bbff1b_0
packaging=23.0=pypi_0
pandas=1.5.3=pypi_0
patsy=0.5.3=pypi_0
pillow=9.4.0=pypi_0
pip=22.3.1=py39haa95532_0
plotly=5.13.0=pypi_0
plotly-express=0.4.1=pypi_0
protobuf=3.20.3=pypi_0
pyarrow=11.0.0=pypi_0
pydeck=0.8.0=pypi_0
pygments=2.14.0=pypi_0
pympler=1.0.1=pypi_0
pyrsistent=0.19.3=pypi_0
python=3.9.16=h6244533_0
python-dateutil=2.8.2=pypi_0
pytz=2022.7.1=pypi_0
pytz-deprecation-shim=0.1.0.post0=pypi_0
requests=2.28.2=pypi_0
rich=13.3.1=pypi_0
scipy=1.10.0=pypi_0
semver=2.13.0=pypi_0
setuptools=65.6.3=py39haa95532_0
six=1.16.0=pypi_0
smmap=5.0.0=pypi_0
sqlite=3.40.1=h2bbff1b_0
statsmodels=0.13.5=pypi_0
streamlit=1.17.0=pypi_0
tenacity=8.1.0=pypi_0
toml=0.10.2=pypi_0
toolz=0.12.0=pypi_0
tornado=6.2=pypi_0
typing-extensions=4.4.0=pypi_0
tzdata=2022.7=pypi_0
tzlocal=4.2=pypi_0
urllib3=1.26.14=pypi_0
validators=0.20.0=pypi_0
vc=14.2=h21ff451_1
vs2015_runtime=14.27.29016=h5e58377_2
watchdog=2.2.1=pypi_0
wheel=0.37.1=pyhd3eb1b0_0
wincertstore=0.2=py39haa95532_2
zipp=3.12.0=pypi_0
